+++ modifier-generics/test/BothPlugins.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE Modifiers #-}
 {-# OPTIONS_GHC -fplugin=Language.Haskell.TH.Modifier.Plugin #-}
 {-# OPTIONS_GHC -fplugin=Generics.Modifier.Plugin #-}
@@ -5,9 +6,14 @@
 
 module BothPlugins (Both (..), BothFields (..)) where
 
-import GHC.Generics (Generic)
+import GHC.Generics qualified as GHC
+import Generics.Modifier (Generic)
 
 %"both"
-data Both = Both deriving (Generic)
+data Both = Both
+  deriving stock GHC.Generic
+  deriving anyclass Generic
 
-data BothFields = BothFields {bothValue %"field" :: Int} deriving (Generic)
+data BothFields = BothFields {bothValue %"field" :: Int}
+  deriving stock GHC.Generic
+  deriving anyclass Generic